Move some Ui primitive types to Engine
