+++ src/Engine/include/OpenLoco/Engine/Ui/Point.hpp
@@ -11,16 +11,4 @@ namespace OpenLoco::Ui
     // Until interop is removed this is a requirement (for global vars mainly)
     static_assert(sizeof(Point) == 4);
     static_assert(sizeof(Point32) == 8);
-
-    struct Size
-    {
-        uint16_t width = 0;
-        uint16_t height = 0;
-
-        constexpr Size(const uint16_t _width, const uint16_t _height)
-            : width(_width)
-            , height(_height)
-        {
-        }
-    };
 }
