+++ src/Engine/include/OpenLoco/Engine/Ui/Rect.hpp
@@ -1,6 +1,8 @@
 #pragma once
 
-#include "UiTypes.hpp"
+#include "Point.hpp"
+#include "Size.hpp"
+
 #include <algorithm>
 #include <cstddef>
