+++ src/OpenLoco/CMakeLists.txt
@@ -452,12 +452,10 @@ set(OLOCO_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Cursor.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Dropdown.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/ProgressBar.h"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Rect.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Screenshot.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/ScrollFlags.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/ScrollView.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/TextInput.h"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/UiTypes.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/WindowManager.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/WindowType.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/Orders.h"
