+++ src/OpenLoco/src/Drawing/SoftwareDrawingContext.h
@@ -4,7 +4,7 @@
 #include "Graphics/Gfx.h"
 #include "Graphics/PaletteMap.h"
 #include "Types.hpp"
-#include "Ui/Rect.h"
+#include <OpenLoco/Engine/Ui/Rect.hpp>
 #include <cstdint>
 
 namespace OpenLoco::Drawing
