+++ src/OpenLoco/src/Input.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#include "Ui/UiTypes.hpp"
 #include "Ui/WindowManager.h"
 #include "Window.h"
 #include <OpenLoco/Core/EnumFlags.hpp>
+#include <OpenLoco/Engine/Ui/Point.hpp>
 
 namespace OpenLoco::Input
 {
